Development

+++ app.py
@@ -1,17 +1,15 @@
 from flask import Flask
 from flask import jsonify, request, make_response, session, send_file, send_from_directory
 from flask_cors import CORS, cross_origin
+from flask_session import Session
 from werkzeug.utils import secure_filename
 from io import BytesIO
 import os
-import sqlite3
+import pymysql
 
 from db import DbUtil
 
-# UPLOAD_FOLDER = '/Users/jacquesdutoit/Documents/vsc/mimer-be'
 UPLOAD_FOLDER = './docs'
-# C:\Users\Jacques\OneDrive - Aesir Systems (Pty) Ltd\Documents\vsc\mimer-be\docs
-# "C:\Users\Jacques\OneDrive - Aesir Systems (Pty) Ltd\Documents\vsc\mimer-be"
 ALLOWED_EXTENSIONS = set(['pdf', 'png'])
 
 app = Flask(__name__)
@@ -20,53 +18,56 @@ app.config['SESSION_PERMANENT'] = True
 app.config['SESSION_TYPE'] = 'filesystem'
 app.config["SESSION_COOKIE_SAMESITE"] = 'None'
 app.config["SESSION_COOKIE_SECURE"] = True
-app.config["SESSION_COOKIE_HTTPONLY"] = False
+app.config["SESSION_COOKIE_HTTPONLY"] = True
 app.config["UPLOAD_FOLDER"] = UPLOAD_FOLDER
 
 db = DbUtil()
+Session(app)
 
-con = sqlite3.connect('mimir.db', check_same_thread=False)
+con = pymysql.connect(host='localhost', user='root', database='mimir')
 cur = con.cursor()
 cur.execute("""
     CREATE TABLE IF NOT EXISTS users (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        name TEXT,
-        surname TEXT,
-        email TEXT,
-        password TEXT
+        id INT(5) PRIMARY KEY AUTO_INCREMENT,
+        name VARCHAR(20),
+        surname VARCHAR(20),
+        email VARCHAR(50) UNIQUE,
+        password VARCHAR(20)
     )
  """)
 cur.execute("""   
     CREATE TABLE IF NOT EXISTS sites (
-        site TEXT PRIMARY KEY,
-        latitude TEXT,
-        longitude TEXT,
-        building TEXT,
-        street TEXT,
-        number TEXT,
-        suburb TEXT,
-        city TEXT,
-        postcode TEXT,
-        province TEXT
+        id INT(5) PRIMARY KEY AUTO_INCREMENT,
+        site VARCHAR(50) UNIQUE,
+        latitude VARCHAR(15),
+        longitude VARCHAR(15),
+        building VARCHAR(30),
+        street VARCHAR(30),
+        number VARCHAR(5),
+        suburb VARCHAR(30),
+        city VARCHAR(30),
+        postcode VARCHAR(10),
+        province VARCHAR(30)
     )
 """)
 cur.execute("""
     CREATE TABLE IF NOT EXISTS circuits (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        vendor TEXT,
-        circuitType TEXT,
-        speed TEXT,
-        circuitNumber TEXT,
-        enni TEXT,
-        vlan TEXT,
-        startDate TEXT,
-        contractTerm TEXT,
-        endDate TEXT,
-        siteA TEXT,
-        siteB TEXT,
-        comments TEXT,
-        status TEXT,
-        doc TEXT
+        id INT(5) PRIMARY KEY AUTO_INCREMENT,
+        vendor VARCHAR(15),
+        circuitType VARCHAR(15),
+        speed VARCHAR(15),
+        circuitNumber VARCHAR(50),
+        enni VARCHAR(15),
+        vlan VARCHAR(15),
+        startDate VARCHAR(15),
+        contractTerm VARCHAR(15),
+        endDate VARCHAR(15),
+        mrc VARCHAR(15),
+        siteA VARCHAR(50),
+        siteB VARCHAR(50),
+        comments VARCHAR(1000),
+        status VARCHAR(15),
+        doc VARCHAR(100)
     )
 """)
 con.close()
@@ -82,10 +83,9 @@ def dict_factory(cursor, row):
 def login():
     obj = request.get_json()
     row = db.get_user_by_email(obj['email'])
-    
+        
     if request.method == 'POST':
         if row:
-            # print(row)
             if obj['password'] == row[4]:
                 session['email'] = row[3]
                 session['fullname'] = row[1] + ' ' + row[2]
@@ -116,8 +116,8 @@ def logout():
         return jsonify({"msg": "You weren't logged in"})
 
 
-@app.route('/register', methods=['POST'])
-@cross_origin(methods=['POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
+@app.route('/register', methods=['GET', 'POST'])
+@cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def register():
     obj = request.get_json()
     # print(obj)
@@ -137,103 +137,92 @@ def register():
                 return jsonify({"msg" : "Welcome to the knowledge of the gods"})
         else:
             return jsonify({"msg": "User already exists"})
+        
+@app.route('/navbar', methods=['GET', 'POST'])
+@cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
+def navbar():
+    if request.method == 'GET':
+        if session:
+            row = db.get_user_by_email(session['email'])
+            # print(row) 
+            res = make_response({'id': row[0], 'email': row[3]})
+            res.status_code = 200
+            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
+            return res
+        elif not session:
+            res = make_response({"error": 'You are not authorized'})
+            res.status_code = 200
+            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
+            return res
    
 @app.route('/circuits', methods=['GET', 'POST'])
 @cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def circuits():
     if request.method == 'GET':
         if session:
-            row = db.get_user_by_email(session['email'])
-            # print(row) 
-            res = make_response({'id': row[0], 'email': row[3]})
-            res.status_code = 200
-            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
-            # print(res.response)
-            return res
-        elif not session:
-            res = make_response({"error": 'You are not authorized'})
-            res.status_code = 200
-            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
-            # print(res.response)
-            return res
-
+            return jsonify({'id': db.get_user_by_email(session['email'])[0], 'email': session['email']})
+        else:
+            return jsonify({"error": 'You are not authorized'}), 401
+        
     if request.method == 'POST':
         obj = request.get_json()
-        # print('This is the object')
-        # print(obj)
-        y = []
-        for key, value in obj.items():
-            if value == "":
-                pass
-            else:
-                value = "'%" + value + "%'"
-                # print(key)
-                # print(value)
-                y = db.search_similar_circuit(key, value)
-        if y:
-            # print(y)
-            return y
-            
-        return jsonify({"error": "No entries found"}), 404
+
+    if not any(obj.values()):
+        return jsonify({"error": "Please enter at least one search parameter"}), 404
+    
+    query = 'SELECT * FROM circuits WHERE '
+    for key, value in obj.items():
+        if value:
+            query += f'{key} LIKE %s AND '
+    query = query.rstrip(' AND ')
+    y = db.search_similar_circuit(query, tuple('%' + value + '%' for value in obj.values() if value))
+    if y:
+        return jsonify(y), 200
+    return jsonify({"error": "No entries found"}), 404
 
 @app.route('/sites', methods=['GET', 'POST'])
 @cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def sites():
     if request.method == 'GET':
         if session:
-            row = db.get_user_by_email(session['email'])
-            # print(row) 
-            res = make_response({'id': row[0], 'email': row[3]})
-            res.status_code = 200
-            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
-            # print(res.response)
-            return res
-        elif not session:
-            res = make_response({"error": 'You are not authorized'})
-            res.status_code = 200
-            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
-            # print(res.response)
-            return res
+            return jsonify({'id': db.get_user_by_email(session['email'])[0], 'email': session['email']})
+        else:
+            return jsonify({"error": 'You are not authorized'}), 401
 
     if request.method == 'POST':
         obj = request.get_json()
-        # print('This is the object')
-        # print(obj)
-        y = []
+
+        if not any(obj.values()):
+            return jsonify({"error": "Please enter at least one search parameter"}), 404
+        
+        query = 'SELECT * FROM sites WHERE '
         for key, value in obj.items():
-            if value == "":
-                pass
-            else:
-                value = "'%" + value + "%'"
-                # print(key)
-                # print(value)
-                y = db.search_similar_site(key, value)
+            if value:
+                query += f'{key} LIKE %s AND '
+        query = query.rstrip(' AND ')
+        y = db.search_similar_site(query, tuple('%' + value + '%' for value in obj.values() if value))
         if y:
-            # print(y)
-            return y
-            
+            return jsonify(y), 200
         return jsonify({"error": "No entries found"}), 404
         
 
-@app.route('/addcircuit', methods=['GET', 'POST'])
-@cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
+@app.route('/circuits/addcircuit', methods=['POST'])
+@cross_origin(methods=['POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def addcircuit():
     if request.method == 'POST':
         obj = request.get_json()
+        x = float(obj['mrc'])
+        obj['mrc'] = "{:.2f}".format(x)
         status = 'Active'
         if obj['doc']:
             doc = obj['doc']
             filename = doc.split('\\')
             filename = filename[2]
-            # print('Original filename: ' + filename)
-            # input()
-            # filename = filename.replace(' ', '_')
             obj['doc'] = secure_filename(filename)
-            # print('Replaced with: ' + filename)
-            # input()
         else:
             filename = 'None'
-        # print(obj)
+            obj['doc'] = filename
+        # print(obj)    
         try:
             db.save_circuit(
                 obj['vendor'],
@@ -244,33 +233,33 @@ def addcircuit():
                 obj['vlan'], 
                 obj['startDate'], 
                 obj['contractTerm'], 
-                obj['endDate'], 
+                obj['endDate'],
+                obj['mrc'], 
                 obj['siteA'],
                 obj['siteB'],
                 obj['comments'],
                 status,
-                filename
+                obj['doc']
+                
             )
             res = make_response({"msg": "Circuit successfully added"})
             res.status_code = 200
             res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
             return res
-        except:   
+        except Exception as e:   
+            print(f"Error: {e}")
             res = make_response({"error": "Unable to save circuit"})
-            res.status_code = 403
-            res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
+            res.status_code = 500  # Use 500 for server errors, not 403
             return res
 
 @app.route('/upload', methods=['POST'])
 @cross_origin(methods=['POST'], supports_credentials=True, origins='http://localhost:3000')
 def upload():
-    # target = os.path.join(UPLOAD_FOLDER, 'docs')
     if not os.path.isdir(UPLOAD_FOLDER):
         os.mkdir(UPLOAD_FOLDER)
     file = request.files['formFile']
     filename = secure_filename(file.filename)
     destination = '/'.join([UPLOAD_FOLDER, filename])
-    # print(filename)
     if filename not in os.listdir(UPLOAD_FOLDER):
         file.save(destination)
     else:
@@ -279,10 +268,9 @@ def upload():
         return res
     res = make_response({"msg": "Document uploaded successfully!"})
     res.status_code = 200
-    # res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
     return res
 
-@app.route('/addsite', methods=['GET', 'POST'])
+@app.route('/sites/addsite', methods=['GET', 'POST'])
 @cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def addsite():
     if request.method == 'POST':
@@ -312,48 +300,44 @@ def addsite():
             res.headers['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'] = True
             return res
         
-@app.route('/viewcircuit/<int:id>', methods=['GET'])
+@app.route('/circuits/viewcircuit/<int:id>', methods=['GET'])
 @cross_origin(methods=['GET'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def view_circuit(id):
     row = db.search_circuit_to_view(id)
+    # print('row found')
     # print(row)
     return row[0]
 
-@app.route('/viewsite/<site>', methods=['GET', 'POST'])
+@app.route('/sites/viewsite/<site>', methods=['GET', 'POST'])
 @cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def view_site(site):
+    row = db.search_site_to_view(site)
     if request.method == 'GET':
-        row = db.search_site_to_view(site)
-        # print(row)
-        return row[0]
+        return row[0] if row else jsonify({"error": "No site found"}), 404
 
     if request.method == 'POST':
-        row = db.search_site_to_view(site)
-        # print('row found')
-        # print(row)
-        db.delete_site(site)
-        return jsonify({"msg": "Deleted!"})
+        if row:
+            db.delete_site(site)
+            return jsonify({"msg": "Deleted!"})
+        return jsonify({"error": "No site found"}), 404
 
-@app.route('/updatecircuit/<int:id>', methods=['POST'])
-@cross_origin(methods=['POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
+@app.route('/circuits/updatecircuit/<int:id>', methods=['GET', 'POST'])
+@cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def update_circuit(id):
+    row = db.search_circuit_to_view(id)[0]
+    if request.method == 'GET':
+        # print(row)
+        return row
+
     if request.method == 'POST':
         obj = request.get_json()
-        # print(obj)
-        if obj['doc']:
+        if 'doc' in obj:
             doc = obj['doc']
-            # print(doc)
-            # input()
-            filename = doc.split('\\')
-            filename = filename[2]
-            # filename = filename.replace(' ', '_')
+            filename = doc.split('\\')[2]
             obj['doc'] = secure_filename(filename)
-            # print(obj['doc'])
         for key, value in obj.items():
-            if key == 'id':
-                pass
-            else:
-                db.update_circuit(key, value, obj['id'])
+            if key != 'id':
+                db.update_circuit(key, value, id)
         return jsonify({"msg": 'Updated'})
     
 @app.route('/download/<int:id>', methods=['GET'])
@@ -362,10 +346,6 @@ def download(id):
     row = db.search_circuit_to_view(id)
     # print(row)
     file = row[0]['doc']
-    # print('FILE: ')
-    # print(file)
-    # print('UPLOAD FOLDER: ')
-    # print(UPLOAD_FOLDER)
     target = '/'.join([UPLOAD_FOLDER, file])
     # print(target)
     if file in os.listdir(UPLOAD_FOLDER):
@@ -378,15 +358,17 @@ def download(id):
 @cross_origin(methods=['GET', 'POST'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
 def get_site():
     obj = request.get_json()
-    if obj != "":
-        obj = "'%" + obj + "%'"
-        # print(obj)
-        y = db.search_sitename(obj)
-        # print(y)
-    else:
-        return jsonify({"msg": "No site found"})
-    return y
+    if obj:
+        query = "SELECT * FROM sites WHERE site LIKE %s"
+        values = (f"%{obj}%",)
+        return db.search_sitename(query, values)
+    return jsonify({"msg": "No site found"})
+
+# @app.route('/test', methods=['GET'])
+# @cross_origin(methods=['GET'], headers=['Content-Type', 'Authorization', 'Access-Control-Allow-Origin'], supports_credentials=True, origins='http://localhost:3000')
+# def test():
+#     return jsonify({"msg": "It works"})
 
 if __name__ == '__main__':
     CORS(app, supports_credentials=True, resource={r"/*": {"origins": "*"}})
-    app.run(debug=True)
+    app.run()
